+++ wgpu-core/src/device/global.rs
@@ -1111,7 +1111,7 @@ impl Global {
                     id: shader.to_trace(),
                     data: file_names,
                     label: desc.label.clone(),
-                    num_workgroups: desc.num_workgroups,
+                    entry_points: desc.entry_points.clone(),
                 });
             };
 
@@ -1405,7 +1405,7 @@ impl Global {
                         Ok(module) => module,
                         Err(e) => break 'error e,
                     };
-                    if module.interface.is_none() {
+                    if module.interface.interface().is_none() {
                         passthrough_stages |= wgt::ShaderStages::VERTEX;
                     }
                     let stage = ResolvedProgrammableStageDescriptor {
@@ -1435,7 +1435,7 @@ impl Global {
                             Ok(module) => module,
                             Err(e) => break 'error e,
                         };
-                        if module.interface.is_none() {
+                        if module.interface.interface().is_none() {
                             passthrough_stages |= wgt::ShaderStages::TASK;
                         }
                         let state = ResolvedProgrammableStageDescriptor {
@@ -1462,7 +1462,7 @@ impl Global {
                         Ok(module) => module,
                         Err(e) => break 'error e,
                     };
-                    if mesh_module.interface.is_none() {
+                    if mesh_module.interface.interface().is_none() {
                         passthrough_stages |= wgt::ShaderStages::VERTEX;
                     }
                     let mesh_stage = ResolvedProgrammableStageDescriptor {
@@ -1493,7 +1493,7 @@ impl Global {
                     Ok(module) => module,
                     Err(e) => break 'error e,
                 };
-                if module.interface.is_none() {
+                if module.interface.interface().is_none() {
                     passthrough_stages |= wgt::ShaderStages::FRAGMENT;
                 }
                 let stage = ResolvedProgrammableStageDescriptor {
@@ -1658,7 +1658,7 @@ impl Global {
                 Ok(module) => module,
                 Err(e) => break 'error e.into(),
             };
-            if module.interface.is_none() && layout.is_none() {
+            if module.interface.interface().is_none() && layout.is_none() {
                 break 'error pipeline::CreateComputePipelineError::Implicit(
                     pipeline::ImplicitLayoutError::Passthrough(wgt::ShaderStages::COMPUTE),
                 );