+++ wgpu-hal/src/metal/mod.rs
@@ -808,7 +808,7 @@ pub enum ShaderModuleSource {
 #[derive(Debug)]
 pub struct PassthroughShader {
     pub library: Retained<ProtocolObject<dyn MTLLibrary>>,
-    pub num_workgroups: (u32, u32, u32),
+    pub num_workgroups: HashMap<String, (u32, u32, u32)>,
 }
 
 unsafe impl Send for PassthroughShader {}